+++ Tokenizers/FixedListTokenizer.cs
@@ -73,7 +73,7 @@ public abstract class FixedListTokenizer : Tokenizer
     /// <param name="ch">The character to check (or begin checking with).</param>
     /// <returns><c>true</c> if this tokenizer accepts the job of parsing the next token
     /// or <c>false</c> if not.</returns>
-    internal override bool CanStart(char ch)
+    protected override bool CanStart(char ch)
     {
         return FindMatch(ch) != null;
     }