+++ Tokenizers/StringTokenizer.cs
@@ -65,7 +65,7 @@ public class StringTokenizer : Tokenizer
     /// <param name="ch">The character to check (or begin checking with).</param>
     /// <returns><c>true</c> if this tokenizer accepts the job of parsing the next token
     /// or <c>false</c> if not.</returns>
-    internal override bool CanStart(char ch)
+    protected override bool CanStart(char ch)
     {
         return IsNext(ch, Bounder);
     }